server/requirements.txt
fastapi 
streamlit 
uvicorn 
pymongo
pydantic
scikit-learn
pandas
pillow

